Renamed filter module; Refactored imports; Added docs
