+++ kraken/src/modules/mod.rs
@@ -5,9 +5,9 @@ pub mod aggregator;
 pub mod attack_results;
 pub mod attacks;
 pub mod cache;
+pub mod filter;
 pub mod oauth;
 pub mod raw_query;
-pub mod syntax;
 pub mod tls;
 pub mod uri;
 mod utc;
