--- kraken/src/modules/syntax/mod.rs
@@ -1,10 +1,0 @@
-//! This module holds all the code for lexing & parsing the filter syntax
-
-pub use self::ast::*;
-pub use self::lexer::*;
-pub use self::parser::ParseError;
-
-mod ast;
-mod lexer;
-mod parser;
-mod sqler;
